Fix missing escaping of newlines and tabs in message
According to the XML spec, when using these in attribute values they should be
escaped to be preserved; otherwise they will be interpreted as an ordinary space.
The xmlEscape function only performs the necessary escaping for text node contents,
so we add a wrapper around it to escape for attribute values.